In employee Details i removed if statement that had no function.
In add record Dialog i removed the Double.parseDouble - as it had no
function.
Refactored the pps Changes.

Created a colour class to hold values for all colours used.

created a miglayout class to hold the miglayout strings - this is a
update  on the method.

After deleting all i removed the dialog that popped up "No employee
Registered" when  click that add button- this pop up will only appear
when the list all employees button is pressed.

Refactored the name of RamdomFile to mangageFile - more to the point.